+++ R/vars-select.R
@@ -15,6 +15,17 @@
 #' * Calls to helper functions are evaluated in the calling context
 #'   only. You can safely and directly refer to local objects.
 #'
+#'
+#' @section Customising error messages:
+#'
+#' For consistency with dplyr, error messages refer to "columns" by
+#' default. This assumes that the variables being selected come from a
+#' data frame. If this is not appropriate for your DSL, you can add an
+#' attribute `vars_type` to the `.vars` vector to specify alternative
+#' names. This must be a character vector of length 2 whose first
+#' component is the singular form and the second is the plural. For
+#' example, `c("variable", "variables")`.
+#'
 #' @param .vars A character vector of existing column names.
 #' @param ...,args Expressions to compute
 #'
@@ -153,7 +164,8 @@ vars_select <- function(.vars, ..., .include = character(), .exclude = character
     first <- ind_list[!is_integerish][[1]]
     first_type <- friendly_type(type_of(first))
     bad_calls(bad,
-      "must resolve to integer column positions, not {first_type}"
+      "must resolve to integer { singular(.vars) } positions, \\
+       not { first_type }"
     )
   }
 
@@ -241,7 +253,10 @@ match_var <- function(chr, table) {
   pos <- match(chr, table)
   if (any(are_na(pos))) {
     chr <- glue::collapse(chr[are_na(pos)], ", ")
-    abort(glue("Strings must match column names. Unkown columns: {chr}"))
+    abort(glue(
+      "Strings must match { singular(table) } names. \\
+       Unknown { plural(table) }: { chr }"
+    ))
   }
   pos
 }